Add initial support on Windows
This adds support using a named pipe on windows. Right now, this only
works on cmd.exe, not Powershell, since Powershell buffers the leftmost
item in a pipeline. This doesn't really work with the current
architecture since we rely on `stdout` being passed to our `stdin` live.

Powershell has a `tee` alternative called `Tee-Object`, so work might be
able to be done here to make `infuser` run like this on Windows.